SP: Fix symbol error caused by UI functions in cgame
These two functions existin SP's exe which is where UI resides there.

I am leaving references to the proper syscalls but since they do not get used when hud is normally loaded in SP (commented out), I think it is best to not call them as it is a global state affecting menus as well.